pyproject.toml
[project]
name = "CondaRTS"
version = "0.0.1"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pygame-ce>=2.5.5",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.13.0",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "ANN",  # flake8-annotations
    "FBT",  # flake8-boolean-trap
    "I",  # isort
    "RET",  # flake8-return
    "SIM",  # flake8-simplify
    "TC",  # flake8-type-checking
    "UP",  # pyupgrade
]
ignore =[
    "ANN001",  # Missing type annotation for function argument
]

[tool.ruff.lint.per-file-ignores]
# Be lenient with original code:
"CondaRTS.py" = [
    "COM812",  # Trailing comma missing - conflicts with ruff format
]
